Rely on double buffering instead of toggling IR interrupt
The PIC code that inspired this routine disables the IR interrupt on
timeout and waits for the caller of the library to re-enable it. This
introduces latency and makes it hard to gauge the state of the signal
when the interrupt service is resumed. Our original code uses a more
optimal solution in double buffering.